Image Upload According to User
I am calling it a day here. Where I'm at is that I've implemented user funcitonalities for the whole app, in that the data is displayed according to the user ID. I have implemented UserID fields to ViewModels and I have created a database table for all the users that only contains the user ID. Other information about the users are being retrieved from the default user table anyway.

I've had a slight problem updating the Venue thumbnails. I have fixed it but I have the same problem with the Events. So tomorrow, this is where I start.

+++ KonneyTM/Controllers/PeopleController.cs
@@ -1,8 +1,10 @@
 ﻿using KonneyTM.DAL;
 using KonneyTM.ViewModels;
+using Microsoft.AspNet.Identity;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Authentication;
 using System.Web;
 using System.Web.Mvc;
 
@@ -12,7 +14,15 @@ namespace KonneyTM.Controllers
     {
         public ActionResult Index()
         {
-            return View(PersonViewModel.GetAllAsOrderedList());
+            if(User.Identity.IsAuthenticated)
+            {
+                var userID = User.Identity.GetUserId();
+                return View(PersonViewModel.GetAll(userID));
+            }
+            else
+            {
+                return View(PersonViewModel.GetAll("demo"));
+            }
         }
 
         public ActionResult Create()
@@ -26,7 +36,16 @@ namespace KonneyTM.Controllers
         {
             if (ModelState.IsValid)
             {
-                personVM.SaveToDB();
+                if(User.Identity.IsAuthenticated)
+                {
+                    var userID = User.Identity.GetUserId();
+                    personVM.SaveToDB(userID);
+                }
+                else
+                {
+                    personVM.SaveToDB("demo");
+                }
+
                 return RedirectToAction("Index");
             }
             return View(personVM);
@@ -35,9 +54,29 @@ namespace KonneyTM.Controllers
         public ActionResult Edit(int id)
         {
             using (var db = new KonneyContext())
-            { 
+            {
                 var personVM = PersonViewModel.FromPerson(db.People.First(p => p.ID == id));
-                return View(personVM);
+
+                if (User.Identity.IsAuthenticated)
+                {
+                    var userID = User.Identity.GetUserId();
+                    if (personVM.UserID == userID)
+                    {
+                        return View(personVM);
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to edit this person.");
+                    }
+                }
+                else if(personVM.UserID == "demo")
+                {
+                    return View(personVM);
+                }
+                else
+                {
+                    throw new Exception("Something went wrong...");
+                }
             }
         }
 
@@ -46,7 +85,28 @@ namespace KonneyTM.Controllers
         {
             if (ModelState.IsValid)
             {
-                personVM.SubmitChanges();
+                if (User.Identity.IsAuthenticated)
+                {
+                    var userID = User.Identity.GetUserId();
+                    if (personVM.UserID == userID)
+                    {
+                        personVM.SubmitChanges(userID);
+                        
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to edit this person.");
+                    }
+                }
+                else if (personVM.UserID == "demo")
+                {
+                    personVM.SubmitChanges("demo");
+                }
+                else
+                {
+                    throw new Exception("Something went wrong...");
+                }
+
                 return RedirectToAction("Index");
             }
             return View(personVM);
@@ -57,7 +117,30 @@ namespace KonneyTM.Controllers
             using (var db = new KonneyContext())
             {
                 var person = db.People.First(p => p.ID == id);
-                db.People.Remove(person);
+
+                if (User.Identity.IsAuthenticated)
+                {
+                    var userID = User.Identity.GetUserId();
+                    if (person.User.ID == userID)
+                    {
+                        db.People.Remove(person);
+
+                    }
+                    else
+                    {
+                        throw new AuthenticationException("You are not authorized to delete this person.");
+                    }
+                }
+                else if (person.User.ID == "demo")
+                {
+                    db.People.Remove(person);
+                    
+                }
+                else
+                {
+                    throw new Exception("Something went wrong...");
+                }
+
                 db.SaveChanges();
             }
             return RedirectToAction("Index");
